third_party/WebKit/LayoutTests/fast/dom/viewport/visualviewportchanged-event-fired-window-resized-expected.txt

This test verifies that that visualviewportchanged event gets fired when the window is resized.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS numCalls is 1
PASS numCalls is 2
PASS successfullyParsed is true

TEST COMPLETE
